Load Illstack data into database
