+++ illstack_helpers.py
@@ -1,0 +1,35 @@
+import numpy as np
+
+NPZ_FILENAME = "sample_illstack.npz"
+
+# List of properties
+global_properties_list = ['M_Crit200', 'R_Crit200', 'GroupFirstSub', 'sfr', 'mstar', 'GroupBHMass', 'GroupBHMdot', 'Group_GasH', 'Group_GasHe', 'Group_GasC', 'Group_GasN', 'Group_GasO', 'Group_GasNe', 'Group_GasMg', 'Group_GasSi', 'Group_GasFe', 'GroupGasMetallicity', 'GroupLen', 'GroupMass', 'GroupNsubs', 'Group_StarH', 'Group_StarHe', 'Group_StarC', 'Group_StarN', 'Group_StarO', 'Group_StarNe', 'Group_StarMg', 'Group_StarSi', 'Group_StarFe', 'GroupStarMetallicity', 'GroupVelx', 'GroupVely', 'GroupVelz', 'GroupWindMass', 'M_Crit500', 'M_Mean200', 'M_TopHat200', 'R_Crit500', 'R_Mean200', 'R_TopHat200']
+
+profile_properties_list = ["val", "n"]
+
+def get_illstack_global_properties():
+    with np.load(NPZ_FILENAME) as f:
+        data = dict(f)
+
+    global_properties_dict = {
+        k : v
+        for k, v in data.items()
+        if k in global_properties_list
+    }
+
+    return global_properties_dict
+
+
+def get_illstack_profile_properties():
+    with np.load(NPZ_FILENAME) as f:
+        data = dict(f)
+
+    radial_bins = data["r"] # List of length 25 (25 == data["nbis"])
+
+    profile_properties_dict = {
+        k : v
+        for k, v in data.items()
+        if k in profile_properties_list
+    }
+
+    return radial_bins, profile_properties_dict
